+++ R/utils-sweave.R
@@ -24,10 +24,10 @@
 #' \code{results=tex/verbatim/hide} are changed to
 #' \code{results='asis'/'markup'/'hide'}; \code{width/height} are changed to
 #' \code{fig.width/fig.height}; \code{prefix.string} is changed to
-#' \code{fig.path}; \code{print/term/strip.white/prefix=TRUE/FALSE} are removed;
-#' most of the character options (e.g. \code{engine} and \code{out.width}) are
-#' quoted; \code{keep.source=TRUE/FALSE} is changed to \code{tidy=FALSE/TRUE}
-#' (note the order of values).
+#' \code{fig.path}; \code{print/term/prefix=TRUE/FALSE} are removed; most of the
+#' character options (e.g. \code{engine} and \code{out.width}) are quoted;
+#' \code{keep.source=TRUE/FALSE} is changed to \code{tidy=FALSE/TRUE} (note the
+#' order of values).
 #'
 #' If a line \code{@@} (it closes a chunk) directly follows a previous
 #' \code{@@}, it is removed; if a line \code{@@} appears before a code chunk and
@@ -138,8 +138,8 @@ fix_sweave = function(x) {
   x = gsub_msg("replacing prefix.string=foo with fig.path='foo'",
                'prefix.string\\s*=\\s*([^,]+)', "fig.path='\\1'", x)
 
-  x = gsub_msg("removing options 'print', 'term', 'stripe.white', 'prefix'",
-               '(print|term|strip.white|prefix)\\s*=\\s*(TRUE|FALSE|T|F)', '', x)
+  x = gsub_msg("removing options 'print', 'term', 'prefix'",
+               '(print|term|prefix)\\s*=\\s*(TRUE|FALSE|T|F)', '', x)
 
   x = gsub_msg('quoting options engine, fig.path, cache.path, fig.keep, fig.show, dev, out.width, out.height, fig.align',
                "(engine|fig\\.path|cache\\.path|fig\\.keep|fig\\.show|dev|out\\.width|out\\.height|fig\\.align)\\s*=\\s*([^,'\"]+)",
@@ -165,7 +165,7 @@ which_sweave = function(x) {
     grep('^<<.*results\\s*=\\s*(tex|verbatim|hide)).*>>=', x),
     grep('^<<.*(fig|pdf|eps|jpeg|png|tikz)\\s*=\\s*(true|false|T|F).*>>=', x),
     grep('^<<.*([, ])(width|height)\\s*=\\s*(\\d+).*>>=', x),
-    grep('^<<.*(keep.source|print|term|strip.white|prefix)\\s*=\\s*(true|false|T|F).*>>=', x)
+    grep('^<<.*(keep.source|print|term|prefix)\\s*=\\s*(true|false|T|F).*>>=', x)
   ))
 }